// frontend/src/app/api/generated/model/patient-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PatientDto { 
    uid?: string;
    no?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    dateOfBirth?: string;
    gender?: string;
    type?: PatientDto.TypeEnum;
    paymentType?: PatientDto.PaymentTypeEnum;
    membershipNo?: string;
    phoneNo?: string;
    address?: string;
    email?: string;
    nationality?: string;
    nationalId?: string;
    passportNo?: string;
    kinFullName?: string;
    kinRelationship?: string;
    kinPhoneNo?: string;
    active?: boolean;
    insurancePlanUid?: string;
    lastVisitAt?: string;
}
export namespace PatientDto {
    export const TypeEnum = {
        Outpatient: 'OUTPATIENT',
        Outsider: 'OUTSIDER',
        Inpatient: 'INPATIENT',
        Deceased: 'DECEASED'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const PaymentTypeEnum = {
        Cash: 'CASH',
        Insurance: 'INSURANCE'
    } as const;
    export type PaymentTypeEnum = typeof PaymentTypeEnum[keyof typeof PaymentTypeEnum];
}